split the 'planning' directory
